+++ types/src/styles/atelier-cave-dark.d.ts
@@ -1,1 +1,0 @@
-export { atelierCaveDark as default } from "./";