+++ types/src/styles/atelier-estuary-dark.d.ts
@@ -1,1 +1,0 @@
-export { atelierEstuaryDark as default } from "./";