+++ types/src/styles/atelier-forest-dark.d.ts
@@ -1,1 +1,0 @@
-export { atelierForestDark as default } from "./";