+++ types/src/styles/atelier-heath-dark.d.ts
@@ -1,1 +1,0 @@
-export { atelierHeathDark as default } from "./";