+++ types/src/styles/atelier-savanna-dark.d.ts
@@ -1,1 +1,0 @@
-export { atelierSavannaDark as default } from "./";