+++ types/src/styles/gruvbox-light.d.ts
@@ -1,1 +1,0 @@
-export { gruvboxLight as default } from "./";