+++ types/src/styles/qtcreator_dark.d.ts
@@ -1,1 +1,0 @@
-export { qtcreator_dark as default } from "./";