+++ types/src/styles/tomorrow-night-eighties.d.ts
@@ -1,1 +1,0 @@
-export { tomorrowNightEighties as default } from "./";